More recipe changes, implemented windmill block checking, added ftbic:no_auto_recipe tag

+++ src/main/java/dev/ftb/mods/ftbic/kubejs/MachineRecipeJS.java
@@ -52,7 +52,7 @@ public class MachineRecipeJS extends RecipeJS {
 		FTBICUtils.listToJson(outputFluids, json, "outputFluids", FluidStackJS::toJson);
 	}
 
-	public MachineRecipeJS processingTime(int p) {
+	public MachineRecipeJS processingTime(double p) {
 		json.addProperty("processingTime", p);
 		save();
 		return this;
